Move forward model tests to own file
